+++ doc/changelog.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* :release:`1.4.1 <9-8-2017>`
+* :bug:`-` Restore default enabled state for Prowl/NMA/Pushbullet notifications
 * :release:`1.4.0 <8-8-2017>`
 * :feature:`-` Added new hook ``prepare_notification`` to process notifications before being sent by the notifications plugin
 * :feature:`662` Improve notifications plugin, add support for email notifications